LayoutTests/fast/dom/select-selectedIndex-expected.txt


PASS mySelect.options.length is 2
PASS mySelect.selectedIndex is 1
PASS mySelect.options.length is 5
PASS mySelect.selectedIndex is 1
PASS mySelect.options.length is 2
PASS mySelect.selectedIndex is 1
PASS mySelect.options.length is 1
PASS mySelect.selectedIndex is -1
PASS mySelect.options.length is 2
PASS mySelect.selectedIndex is 1
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
PASS mySelect.options.length is 0
PASS mySelect.selectedIndex is -1
PASS mySelect.options.length is 2
PASS mySelect.selectedIndex is 1
PASS mySelect.options.length is 11
PASS mySelect.selectedIndex is 10
PASS mySelect.options.length is 11
PASS mySelect.selectedIndex is 10
PASS mySelect.options.length is 10
PASS mySelect.selectedIndex is -1
PASS mySelect.options.length is 10
PASS mySelect.selectedIndex is -1
PASS mySelect.options.length is 10
PASS mySelect.selectedIndex is -1
